Overhaul of the ipanema scheduling class.
The ioctl() is no longer supported. To use an ipanema policy, you must use the
sched_setattr() syscall or use ipastart (located in tools/ipanema).

The /proc/ipanema_policies file now takes the following syntax in input:
    <cpulist>:<module>
For example, if you have inserted a module with a policy called 'Simple', use:
    echo "0-2,5:Simple" > /proc/ipanema_policies
to create an instance of the 'Simple' policy on cpus 0, 1, 2 and 5. The wildcard
'*' is not valid anymore.

For now, each module can be instanciated only once as a policy (still working on
allowing multiple instances of a module).

Changing the cpulist associated with a policy works, when adding/removing cpus,
the policy sees cpus plugged/unplugged and calls its core_entry/exit() events
accordingly.