Fixed ipanema policies and modules handling (mainly removal)
Policies are now stored in a list (struct list_head) and have a refcount. Every
time a task registers to a policy (by setscheduler() or fork()), it increments
the refcount. Every time a task leaves an ipanema_policy (setscheduler() or
exit()), it decrements the refcount. A policy can be removed only if the
refcount is equal to 1 (ie. no task is using it).

For ipanema modules, each time a policy instance is created, it takes a
reference to the kernel module (struct module, not the struct ipanema_module)
containing the policy's handlers. This means that removing the module while a
policy is instantiated with this module is prevented before entering the
module's cleanup function. When a policy instance is removed, it releases the
reference to the kernel module to allow module removal if not in use by any
policy.