ipanema: refactored topology export to policies to expose hw info with flags
The exported topology is now built using the hierarchy built by Linux with the
sched_domains, in order to get hardware information on the domains. For now, we
export 3 flags: DOMAIN_SMT (cores share computing hardware), DOMAIN_CACHE (cores
share a cache) and DOMAIN_NUMA (memory accesses are not uniform for all cores).