ipanema: removed an erroneous warning
This warning was printed when pick_next_task_ipanema() was called while
ipanema_current was not NULL. This should not happen since an ipanema event
such as block(), yield() or terminate() should have set it to NULL. However,
when a task releases the CPU by calling schedule() with a state set to something
different than TASK_RUNNING (TASK_INTERRUPTIBLE for example), __schedule()
checks if signals are pending. If it is the case, the task is set to
TASK_RUNNING and kept in the queue to handle them. When they are handled, it
will be dequeued, and the ipanema event will be called accordingly, setting
ipanema_current to NULL.

We keep the warning on if task->state != TASK_RUNNING, in case there is
something else triggering it.