ipanema: bug fix : tasks are no longer misplaced on the wrong CPU
Fixed a bug that caused an ipanema task to have a discrepancy in its metadata:
the cpu field of the task_struct and the cpu field of its struct ipanema_rq were
different. This happened because of CFS doind NUMA balancing. Somehow, CFS
sometimes chose a task that was not using the fair sched class, and swapped it
with a task using the fair sched class. This caused the ipanema thread to be
moved to a new CPU, without properly saying to the ipanema sched class to do
something particular to handle this case (that should never happen by the way).

Our (dirty) fix consists in checking if the task chosen by CFS in
task_numa_migrate() is using the ipanema sched class. If this is the case, we
fallback to the behaviour used when no task was chosen.