Ipanema policies can now be managed with cgroups. The insertion API was also simplified.
When an ipanema module is inserted, it is now instantly instantiated on all CPUs. Currently loaded policies can be seen in `/proc/ipanema/policies`. Processes can still use the setscheduler syscall to use an ipanema policy. The novelty here is that you can create an ipanema cgroup bound to a specific policy. Processes in this cgroup will automatically use this policy. In order to assign a policy to an ipanema cgroup, you must write the policy id (located in `/proc/ipanema/policies`) in the `ipanema.policy_id` file. If -1 is written to this file, the cgroup reverts to the fair scheduling class.